SpecialDisplayNotificationsConfiguration: Show conditional options
Instead of two tables for "Existing users" and "New users", we can
simply display a table for each possible combination of conditions –
there probably won't ever be that many of them (and currently there
are still exactly two, corresponding to the old tables).

It would be nice to explain the conditions in human terms, but we can
simply display a serialization of how they're defined in the code.

Given that this is a feature intended for developers and power users,
I think this is acceptable.

Bug: T357219
Change-Id: I4db9ef8a9a5a154e84034140647097c3729ed87d