Simplify Special:Notifications layout using flexbox
Compared to the previous approach using table layouts, it is much
easier to control which elements are allowed to extend to maximum
available width, and to provide alternative layouts for mobile.
There should be no visual changes.

It doesn't require using `white-space: nowrap` to prevent elements
from being squished, which will help with T237244.

(These styles are only loaded on Special:Notifications, they don't
apply to the flyout.)

Bug: T237244
Change-Id: I09e692e427f61fabd53b3824a1ed7e4f8e76fad4