Allow sidebar to be slightly smaller
This reduces the sidebar to 15em using flexbox when required. This
keeps the layout as columns just slightly longer, when compressing
which especially on vector-2022 will happen quite a bit.

This required making the lines in the optionwidget flex as well,
so that the ellipsis doesn't depend on a fixed value any more.

Bug: T337954
Change-Id: I8aadb1799da33161abc703344e9ab90bb747c604