Simplify "my user talk page" to "my talk page"
I believe that saying "my talk page" is canon and really more than
enough. Even less confusing, especially when compared with the other
messages that talk about "my user page".

An additional motivation here is to make this easier to localize.
This turns out to be especially relevant in languages like German
where "user" needs to be gendered. We can avoid this here entirely.

Change-Id: I6588032d2b54e3729200c3342d668fc445f46312