Replace "the user talk page of" with "the talk page of user"
The main motivation here is to make this easier to localize. For
example, in German the word "user" needs to be gendered depending
on the user's preference. But the user name appears only much later
in the sentence, after "of". It appears like this can easily be
improved and even the original English message reads better then,
doesn't it?

Change-Id: I055481edbd572334bd2fffdcad9da67376ca5464