Provide a generic way to pass recipients along with event
Event definition requires developers to specify `user-locators`.
Standard way to pass recipients is to use specific locators
or pass them in the extra array and use locateFromEventExtra()
locator. Different extensions pass recipients in different way (under
recipients, reverted-user-id, thanked-user-id, to-user-id, etc).

With new MediaWiki Notifications system we need a generic way to pass
recipients, so there won't be a need to register custom user locator
with every single event.

In majority of cases we know the recipients when creating the event.
We should use the `extra` array as it is already present, and add a
default `recipients` key, that when specified, will contain list of
user_ids. When `user-locators` event type definition is specified,
use it, otherwise use the `locateFromEventExtra('recipients')` to
retrieve list of recipients.

In case of `emailuser` we need store both `to-user-id` and `recipients`
so we can handle transition from existing locator to new one.

Bug: T385050
Change-Id: I9aa40a9362f269069a298d3e1e1a576982e9069c