feat: default recipient locator should be always present
In I9aa40a9362f269069a298d3e1e1a576982e9069c we introduced a
default UserLocator that would be returned when no user locators
is defined for specific Event. This causes an extra effort during
migration, as per short period of time, we need to support both
locators (the old one being deprecated) and the new one.

The scenario follows:
 - code sends Event that uses old locator
 - deployment happens that removes the old locator and uses the
 `recipients` key
 - code picks up the old Event, because the old locator is removed
 it tries to fetch information from `extra['recipients']`, but this
 Event doesn't have it yet (was triggered before deployment).

As a solution, lets inject the default user locator every time. This
doesn't cause any harm, as if `recipients` is not set, it won't fail,
only ignore it.

Bug: T385050
Change-Id: Iac57eb26a811b53c5b85e52ada25b2ec2a0c3f3b