Hooks: Deduplicate alert/message badge code, allow combined badge
The code to generate the alert and message badges was identical,
except for constant strings, and accidental drift due to code
duplication. Generate them in a loop instead.

Also add a combined badge, handled by the same code, which is
surprisingly easy (the backend already supported this). Enable this
combined badge for the mobile Minerva skin only, replacing the feature
previously implemented hackily in the skin. This works without any
changes to Minerva (although it can be cleaned up later), and is
improved by setting correct hover tooltips and screenreader labels.

Change-Id: If4ae4376621a68521d7a9a92d7d3b9d9929bac73