Add logging to help figure unserialization issues
Somehow, we get an error message in the logs:

> PHP Warning: Class __PHP_Incomplete_Class has no unserializer

However, the unserialization technically completes, so the previous
logging in the exception handler does not trigger.

It is assumed that the over unseralization of the array succeeds, but it
contains one key that points to something that is now unserializable,
maybe because a class-alias has been removed. However, that means that
we now have to walk the entire array to find the few instances of this
going wrong.

Bug: T388725
Change-Id: I5b7b9aa336c1a4a4e45464e52bfc67f3fc1e18e0
(cherry picked from commit e0bdef44e098d84bfe1899011c3eaea3af89b636)