Allow event_extra to be JSON
We're updating the event extra serialization method to JSON. This is
a first part, making sure code can understand and deserialize json
string from `event_extra` column.
The support of reading JSON has to be merged and deployed first. We
cannot do this in one go, as in possible case of rollback, the DB
will have the `event_extra` serialized with JSONCodec but code wouldn't
know how to handle this and deserializing those events would fail.
In order to prevent such failures, first we need to push the reading
part, and after couple weeks push the writing part.

Bug: T325703
Change-Id: Iafc7ba506c5082f14cc3e29293f9144528e49942