Remove some more unused code for replacing the badge icons
The .$element nodes in BadgeLinkWidget and NotificationBadgeWidget
are not used any more, as they instead use the $badge provided in
config, so remove the code that sets them up and rename the classes
from Widget to Controller.

Fix updating the badge accessibility label when a notification
is marked as read, accidentally broken in the previous change.

Follow-up to cb53ea80e5b64331639780b48f6ff4966542b5df.

Bug: T343838
Change-Id: I231ab15f12028a88bce87433fa53f7c312f0d8da