tests: Drop echo.suite.xml
This was last touched in I75e96cc1e51a77, in 2014. It doesn't appear
anywhere in codesearch. I'm not sure if it still works, but it surely
won't work once T345481 is done.

More generally, we want all extensions to work with the standard PHPUnit
setup. It is already complicated enough as it is, we don't need to add
special extension configs to the mix.

Change-Id: Idd8aa09ebb7d8505650eda3edc430aacbc224053