Allow extension notifications to use pagelink mutes
Echo has a setting to opt out of page-linked notifications for specific
pages, implemented using a check that hard-coded that notification type
ID.

Some extensions, like Wikibase, have notifications that are effectively
other variants of the page-linked notification provided in Echo itself.
This patch changes that hard-coded notification type check to check a
$wgEchoNotifications attribute instead, so that those extensions can opt
in for using that list.

Change-Id: Iec0edd9db7df86f70b5c3969f5bce0031d062234